Remove some useless cast.

+++ src/gui/button.cpp
@@ -41,8 +41,8 @@ bool Button::mouseHovering()
     if ((_flags & Disabled) == Disabled)
         return false;
 
-    if ((int)mousePosition().x >= left() && (int)mousePosition().x <= left() + width() &&
-        (int)mousePosition().y >= top() && (int)mousePosition().y <= top() + height())
+    if (mousePosition().x >= left() && mousePosition().x <= left() + width() &&
+        mousePosition().y >= top() && mousePosition().y <= top() + height())
     {
         _flags |= MouseOver;
         return true;
